# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/admin/audit/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Root can query audit logs of all users with pagination and time range filter.\nroot 用户可分页查询全部审计日志，并支持时间范围过滤。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audit"
                ],
                "summary": "List audit logs (admin) / 查询审计日志（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page / 页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "page size / 每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "start time (RFC3339 or YYYY-MM-DD) / 开始时间",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end time (RFC3339 or YYYY-MM-DD) / 结束时间",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "filter by user_id / 按用户ID过滤（可选）",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-api_AuditLogPage"
                        }
                    }
                }
            }
        },
...
        "/api/v1/auth/login": {
            "post": {
                "description": "Create a web session token with sliding idle timeout.\n创建带滑动空闲超时的 Web 会话 Token。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Web login / Web 登录",
                "parameters": [
                    {
                        "description": "login request / 登录请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.LoginWebRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-api_LoginWebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revoke current token (web session or api token).\n撤销当前 Token（Web 会话或 API Token）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout / 注销",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/token": {
            "post": {
                "description": "Create a permanent API token for subsequent authenticated API calls.\n通过用户名密码创建永久 API Token，用于后续需要鉴权的 API。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Create API token by password / 通过密码创建 API Token",
                "parameters": [
                    {
                        "description": "request / 请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateAPITokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-api_CreateAPITokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.AuditLogItem": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "resource": {
                    "type": "string"
                },
                "user_agent": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "api.AuditLogPage": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.AuditLogItem"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "api.CreateAPITokenRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "api.CreateAPITokenResponse": {
            "type": "object",
            "properties": {
                "jti": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "api.LoginWebRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "api.LoginWebResponse": {
            "type": "object",
            "properties": {
                "idle_timeout_seconds": {
                    "type": "integer"
                },
                "jti": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "response.Envelope-any": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.ErrorCode"
                },
                "data": {},
                "message": {
                    "description": "English message / 英文信息",
                    "type": "string"
                }
            }
        },
...
        "response.Envelope-api_AuditLogPage": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.ErrorCode"
                },
                "data": {
                    "$ref": "#/definitions/api.AuditLogPage"
                },
                "message": {
                    "description": "English message / 英文信息",
                    "type": "string"
                }
            }
        },
        "response.Envelope-api_CreateAPITokenResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.ErrorCode"
                },
                "data": {
                    "$ref": "#/definitions/api.CreateAPITokenResponse"
                },
                "message": {
                    "description": "English message / 英文信息",
                    "type": "string"
                }
            }
        },
        "response.Envelope-api_LoginWebResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.ErrorCode"
                },
                "data": {
                    "$ref": "#/definitions/api.LoginWebResponse"
                },
                "message": {
                    "description": "English message / 英文信息",
                    "type": "string"
                }
            }
        },
...
        "response.ErrorCode": {
            "type": "integer",
            "enum": [
                0,
                10000,
                10001,
                10002,
                10003,
                10004,
                10005
            ],
            "x-enum-varnames": [
                "CodeOK",
                "CodeBadRequest",
                "CodeUnauthorized",
                "CodeForbidden",
                "CodeNotFound",
                "CodeConflict",
                "CodeInternal"
            ]
        }
    }
}